+++ fabric/src/main/java/dev/nyon/magnetic/mixins/blocks/PumpkinBlockMixin.java
@@ -1,0 +1,43 @@
+package dev.nyon.magnetic.mixins.blocks;
+
+import com.llamalad7.mixinextras.injector.v2.WrapWithCondition;
+import dev.nyon.magnetic.utils.MixinHelper;
+import net.minecraft.core.BlockPos;
+import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Items;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.PumpkinBlock;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.phys.BlockHitResult;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+
+@Mixin(PumpkinBlock.class)
+public class PumpkinBlockMixin {
+
+    @WrapWithCondition(
+        method = "useItemOn",
+        at = @At(
+            value = "INVOKE",
+            target = "Lnet/minecraft/world/level/Level;addFreshEntity(Lnet/minecraft/world/entity/Entity;)Z"
+        )
+    )
+    private boolean doNotSpawnEntityWhenMagnetic(
+        Level instance,
+        Entity entity,
+        ItemStack stack,
+        BlockState state,
+        Level world,
+        BlockPos pos,
+        Player player,
+        InteractionHand hand,
+        BlockHitResult hitResult
+    ) {
+        if (!(player instanceof ServerPlayer serverPlayer)) return true;
+        return MixinHelper.wrapWithConditionPlayerItemSingle(serverPlayer, new ItemStack(Items.PUMPKIN_SEEDS, 4));
+    }
+}